+++ test/e2e/knative_serving_test.go
@@ -10,20 +10,15 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	pkgTest "knative.dev/pkg/test"
 	servingoperatorv1alpha1 "knative.dev/serving-operator/pkg/apis/serving/v1alpha1"
 )
 
 const (
 	knativeServing                = "knative-serving"
 	testNamespace                 = "serverless-tests"
-	testNamespace2                = "serverless-tests2"
 	image                         = "gcr.io/knative-samples/helloworld-go"
 	proxyImage                    = "gcr.io/knative-samples/autoscale-go:0.1"
 	helloworldService             = "helloworld-go"
-	helloworldService2            = "helloworld-go2"
-	kubeHelloworldService         = "kube-helloworld-go"
-	helloworldText                = "Hello World!"
 	proxyHelloworldServiceSuccess = "proxy-helloworld-go-success"
 	proxyHelloworldService        = "proxy-helloworld-go"
 	httpProxy                     = "HTTP_PROXY"
@@ -79,14 +74,6 @@ func TestKnativeServing(t *testing.T) {
 		}
 	})
 
-	t.Run("user permissions", func(t *testing.T) {
-		testUserPermissions(t)
-	})
-
-	t.Run("deploy knative and kubernetes service in same namespace", func(t *testing.T) {
-		testKnativeVersusKubeServicesInOneNamespace(t, caCtx)
-	})
-
 	t.Run("update global proxy and verify calls goes through proxy server", func(t *testing.T) {
 		t.Skip("SRKVS-462: This test needs thorough hardening")
 		testKnativeServingForGlobalProxy(t, caCtx)
@@ -117,212 +104,6 @@ func TestKnativeServing(t *testing.T) {
 			t.Fatalf("Operators still running: %v", err)
 		}
 	})
-}
-
-func testKnativeVersusKubeServicesInOneNamespace(t *testing.T, caCtx *test.Context) {
-	//Create deployment
-	err := test.CreateDeployment(caCtx, kubeHelloworldService, testNamespace2, image)
-	if err != nil {
-		t.Fatal("Deployment not created", err)
-	}
-	// Deploy plain Kube service
-	svc, err := test.CreateKubeService(caCtx, kubeHelloworldService, testNamespace2)
-	if err != nil {
-		t.Fatal("Kubernetes service not created", err)
-	}
-	route, err := test.WithRouteForServiceReady(caCtx, svc.Name, testNamespace2)
-	if err != nil {
-		t.Fatal("Failed to create route for service", svc.Name, err)
-	}
-	kubeServiceURL := "http://" + route.Status.Ingress[0].Host
-
-	// Check Kube service responds
-	waitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
-
-	// Deploy Knative service in the same namespace
-	ksvc, err := test.WithServiceReady(caCtx, helloworldService, testNamespace2, image)
-	if err != nil {
-		t.Fatal("Knative Service not ready", err)
-	}
-
-	// Check that both services respond
-	waitForRouteServingText(t, caCtx, ksvc.Status.URL.Host, helloworldText)
-	waitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
-
-	// Delete Knative service
-	caCtx.Clients.Serving.ServingV1beta1().Services(testNamespace2).Delete(ksvc.Name, &metav1.DeleteOptions{})
-
-	// Check that Kube service still responds
-	waitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
-
-	// Remove the Kube service
-	caCtx.Clients.Route.Routes(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-	caCtx.Clients.Kube.CoreV1().Services(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-	caCtx.Clients.Kube.AppsV1().Deployments(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-
-	// Deploy Knative service in the namespace first
-	ksvc, err = test.WithServiceReady(caCtx, helloworldService2, testNamespace2, image)
-	if err != nil {
-		t.Fatal("Knative Service not ready", err)
-	}
-
-	// Check that Knative service responds
-	waitForRouteServingText(t, caCtx, ksvc.Status.URL.Host, helloworldText)
-
-	//Create deployment
-	err = test.CreateDeployment(caCtx, kubeHelloworldService, testNamespace2, image)
-	if err != nil {
-		t.Fatal("Deployment not created", err)
-	}
-	// Deploy plain Kube service
-	svc, err = test.CreateKubeService(caCtx, kubeHelloworldService, testNamespace2)
-	if err != nil {
-		t.Fatal("Kubernetes service not created", err)
-	}
-	route, err = test.WithRouteForServiceReady(caCtx, svc.Name, testNamespace2)
-	if err != nil {
-		t.Fatal("Failed to create route for service", svc.Name, err)
-	}
-	kubeServiceURL = "http://" + route.Status.Ingress[0].Host
-
-	// Check that both services respond
-	waitForRouteServingText(t, caCtx, ksvc.Status.URL.Host, helloworldText)
-	waitForRouteServingText(t, caCtx, kubeServiceURL, helloworldText)
-
-	// Remove the Kube service
-	caCtx.Clients.Route.Routes(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-	caCtx.Clients.Kube.CoreV1().Services(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-	caCtx.Clients.Kube.AppsV1().Deployments(testNamespace2).Delete(svc.Name, &metav1.DeleteOptions{})
-
-	// Check that Knative service still responds
-	waitForRouteServingText(t, caCtx, ksvc.Status.URL.Host, helloworldText)
-
-	// Delete the Knative service
-	caCtx.Clients.Serving.ServingV1beta1().Services(testNamespace2).Delete(ksvc.Name, &metav1.DeleteOptions{})
-}
-
-func testUserPermissions(t *testing.T) {
-	paCtx := test.SetupProjectAdmin(t)
-	editCtx := test.SetupEdit(t)
-	viewCtx := test.SetupView(t)
-	test.CleanupOnInterrupt(t, func() { test.CleanupAll(t, paCtx, editCtx, viewCtx) })
-	defer test.CleanupAll(t, paCtx, editCtx, viewCtx)
-
-	tests := []struct {
-		name        string
-		userContext *test.Context
-		operation   func(context *test.Context) error
-		wantErrStr  string
-	}{{
-		name: "user with view role can get",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).Get(helloworldService, metav1.GetOptions{})
-			return err
-		},
-		userContext: viewCtx,
-	}, {
-		name: "user with view role can list",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).List(metav1.ListOptions{})
-			return err
-		},
-		userContext: viewCtx,
-	}, {
-		name: "user with view role cannot create",
-		operation: func(c *test.Context) error {
-			_, err := test.CreateService(c, "userview-service", testNamespace, image)
-			return err
-		},
-		userContext: viewCtx,
-		wantErrStr:  "is forbidden",
-	}, {
-		name: "user with view role cannot delete",
-		operation: func(c *test.Context) error {
-			return c.Clients.Serving.ServingV1beta1().Services(testNamespace).Delete(helloworldService, &metav1.DeleteOptions{})
-		},
-		userContext: viewCtx,
-		wantErrStr:  "is forbidden",
-	}, {
-		name: "user with project admin role can get",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).Get(helloworldService, metav1.GetOptions{})
-			return err
-		},
-		userContext: paCtx,
-	}, {
-		name: "user with project admin role can list",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).List(metav1.ListOptions{})
-			return err
-		},
-		userContext: paCtx,
-	}, {
-		name: "user with project admin role can create",
-		operation: func(c *test.Context) error {
-			_, err := test.CreateService(c, "projectadmin-service", testNamespace, image)
-			return err
-		},
-		userContext: paCtx,
-	}, {
-		name: "user with project admin role can delete",
-		operation: func(c *test.Context) error {
-			return c.Clients.Serving.ServingV1beta1().Services(testNamespace).Delete("projectadmin-service", &metav1.DeleteOptions{})
-		},
-		userContext: paCtx,
-	}, {
-		name: "user with edit role can get",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).Get(helloworldService, metav1.GetOptions{})
-			return err
-		},
-		userContext: editCtx,
-	}, {
-		name: "user with edit role can list",
-		operation: func(c *test.Context) error {
-			_, err := c.Clients.Serving.ServingV1beta1().Services(testNamespace).List(metav1.ListOptions{})
-			return err
-		},
-		userContext: editCtx,
-	}, {
-		name: "user with edit role can create",
-		operation: func(c *test.Context) error {
-			_, err := test.CreateService(c, "useredit-service", testNamespace, image)
-			return err
-		},
-		userContext: editCtx,
-	}, {
-		name: "user with edit role can delete",
-		operation: func(c *test.Context) error {
-			return c.Clients.Serving.ServingV1beta1().Services(testNamespace).Delete("useredit-service", &metav1.DeleteOptions{})
-
-		},
-		userContext: editCtx,
-	}}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			err := test.operation(test.userContext)
-			if (err != nil) != (test.wantErrStr != "") {
-				t.Errorf("User with role %s has unexpected behavior on knative services. Error thrown: %v, error expected: %t", test.userContext.Name, err, (test.wantErrStr != ""))
-			}
-			if err != nil && !strings.Contains(err.Error(), test.wantErrStr) {
-				t.Errorf("Unexpected error for user with role %s: %v", test.userContext.Name, err)
-			}
-		})
-	}
-}
-
-func waitForRouteServingText(t *testing.T, caCtx *test.Context, routeDomain, expectedText string) {
-	t.Helper()
-	if _, err := pkgTest.WaitForEndpointState(
-		&pkgTest.KubeClient{Kube: caCtx.Clients.Kube},
-		t.Logf,
-		routeDomain,
-		pkgTest.EventuallyMatchesBody(expectedText),
-		"WaitForRouteToServeText",
-		true); err != nil {
-		t.Fatalf("The Route at domain %s didn't serve the expected text \"%s\": %v", routeDomain, expectedText, err)
-	}
 }
 
 func testKnativeServingForGlobalProxy(t *testing.T, caCtx *test.Context) {
